Python Skript zur Aufnahme der Camera Objektlisten erstellt, Launchfile hinzugefügt, Argumente zum Starten der Bagfile Record Skripte hinzugefügt
